:recycle: Large-scale refactor.
This comes with a lot of changes, but mostly this converts spacetar from a one-file cli tool into a complete Python package, meant for being used in Python programs and on the terminal. There is currently no documentation for anything, but that will be fixed in future commits. A lot of code clean-up is also required I think.