:bug: Fix lazy loading error.
This commit fixes an error encountered when the user searches for a source/telescope in the corresponding table and attempts to load the `molecules` attribute. Since SQLAlchemy attempted to lazy load it, this lead to an error (since it was not bound to a any `Session`). Adding a backref to the loading strategy ensures that the `molecules` attribute is also loaded correctly.