:sparkles: Added new methods to Molecule class.
This commit uses the chimie module to add some new methods to the Molecule class. These let the user know the composition, charge, number of atoms, molecular mass, and degree of unsaturation of the molecule, all using the formula string! This is thanks to a simple parser of chemical formulas that is built into the chimie module.