:boom: Change how the search APi works.
This commit introduces the Results class. I finally decided to bite the bullet and subclass list for this, adding a few key features that will allow me to perform some SQL-like searches on non-SQL parameters (such as properties).

This will allow the introduction of additional features, like querying the table by the most or least of certain parameters, such as the discovery year or the number of detections.

The tests have been changed accordingly.